Align test names with master
