+++ src/NServiceBus.Transport.SQS.AcceptanceTests/Sending/When_using_large_message_with_kms_encrypted_bucket.cs
@@ -1,15 +1,11 @@
-namespace NServiceBus.AcceptanceTests.Sending
+﻿namespace NServiceBus.AcceptanceTests.Sending
 {
-    using System;
-    using System.Security.Cryptography;
     using System.Threading.Tasks;
     using AcceptanceTesting;
-    using Amazon.S3;
-    using Amazon.S3.Model;
     using EndpointTemplates;
     using NUnit.Framework;
 
-    public class Sending_large_message_using_customer_provided_aes : NServiceBusAcceptanceTest
+    public class When_using_large_message_with_kms_encrypted_bucket : NServiceBusAcceptanceTest
     {
         [Test]
         public async Task Should_receive_message()
@@ -27,22 +23,13 @@ namespace NServiceBus.AcceptanceTests.Sending
             Assert.AreEqual(payloadToSend, context.ReceivedPayload, "The large payload should be handled correctly using the kms encrypted S3 bucket");
 
             var s3Client = SqsTransportExtensions.CreateS3Client();
-            var getObjectResponse = await s3Client.GetObjectAsync(new GetObjectRequest
-            {
-                BucketName = BucketName,
-                Key = $"{SqsTransportExtensions.S3Prefix}/{context.MessageId}",
-                ServerSideEncryptionCustomerMethod = ServerSideEncryptionCustomerMethod.AES256,
-                ServerSideEncryptionCustomerProvidedKey = Base64Key,
-            });
 
-            Assert.AreEqual(ServerSideEncryptionCustomerMethod.AES256, getObjectResponse.ServerSideEncryptionCustomerMethod);
-            Assert.IsNull(getObjectResponse.ServerSideEncryptionMethod);
+            Assert.DoesNotThrowAsync(async () => await s3Client.GetObjectAsync(BucketName, $"{SqsTransportExtensions.S3Prefix}/{context.MessageId}"));
         }
 
         const int PayloadSize = 150 * 1024;
 
         static string BucketName;
-        static string Base64Key;
 
         public class Context : ScenarioContext
         {
@@ -58,15 +45,9 @@ namespace NServiceBus.AcceptanceTests.Sending
                 {
                     var transportConfig = c.UseTransport<SqsTransport>();
 
-                    BucketName = $"{SqsTransportExtensions.S3BucketName}";
+                    BucketName = $"{SqsTransportExtensions.S3BucketName}.kms";
 
-                    var aesEncryption = Aes.Create();
-                    aesEncryption.KeySize = 256;
-                    aesEncryption.GenerateKey();
-                    Base64Key = Convert.ToBase64String(aesEncryption.Key);
-
-                    var s3Config = transportConfig.S3(BucketName, SqsTransportExtensions.S3Prefix);
-                    s3Config.ServerSideCustomerEncryption(ServerSideEncryptionCustomerMethod.AES256, Base64Key);
+                    transportConfig.S3(BucketName, SqsTransportExtensions.S3Prefix);
                 });
             }
