+++ src/NServiceBus.Transport.SQS.AcceptanceTests/Sending/When_using_oversized_message_without_bucket_configured.cs
@@ -7,7 +7,7 @@
     using NUnit.Framework;
     using Transport.SQS.Configure;
 
-    public class Sending_oversized_message_without_bucket_configured : NServiceBusAcceptanceTest
+    public class When_using_oversized_message_without_bucket_configured : NServiceBusAcceptanceTest
     {
         [Test]
         public void Should_fail()
